Improve bot->block collisions

+++ block.cpp
@@ -2,7 +2,7 @@
 #include "screen.h"
 #include "drawing.h"
 
-Block::Block(int id, int typeCode, int x, int y){
+Block::Block(int id, int typeCode, double x, double y){
 	this->id = id;
 	this->typeCode = typeCode;
 	this->x = x;
